step2 기능을 MVC 패턴으로 구현
